test: split test_creation tests into language specific files
